Add files via upload
This is the reinforcement learning agent class used for learning the state-state successor representation (without Dyna). 
This is for tasks where the agent is not taking any actions, but is taken along a path passively, such as experiment 1 in Momennejad et al. 2017, Nat Hum Beh, or Schapiro et al. 2013. It is similar to statistical learning.